[language] clean up use of gas_schedule::CostStrategy
The diem-vm code has been inconsistent about how it uses CostStrategy values.
In some cases where it wants to use the VM with metering disabled, it
constructs a new `CostStrategy::system`, but in other cases, it disables
metering in an existing CostStrategy. But, it’s more complicated than that --
sometimes when creating a new `system` CostStrategy, it passes the remaining
gas balance from the previous CostStrategy into the new one, essentially
using the CostStrategy as a way to pass the remaining gas balance. This code
also pointlessly passes around `gas_schedule` arguments that are then used
to create `system` CostStrategy values where the gas schedule is ignored.

We should be consistent about this: when we want to run an epilogue function
with metering disabled, we should either consistently create a new `system`
CostStrategy or disable metering on the previous CostStrategy. After trying
it both ways, I think it works better to use one CostStrategy and disable
metering for epilogue functions. This also works well as a way to track the
remaining gas balance.

Also, because `CostStrategy::system` constructs a CostStrategy where metering
is always disabled, there’s no need to specify a gas schedule and the amount
of gas remaining. We can simplify all the places where this is used by
removing those parameters.